Functional WsProvider
