+++ packages/api-provider/src/http/index.js
@@ -4,24 +4,16 @@
 // @flow
 
 import type { ProviderInterface, ProviderInterface$Callback } from '../types';
-import type { HttpState } from './types';
 
 require('./polyfill');
 
-const assert = require('@polkadot/util/assert');
-
-const coder = require('../jsonRpcCoder');
 const send = require('./send');
 const subscribe = require('./subscribe');
+const state = require('./state');
 const unsubscribe = require('./unsubscribe');
 
 module.exports = function httpProvider (endpoint: string): ProviderInterface {
-  const self: HttpState = {
-    coder: coder(),
-    endpoint
-  };
-
-  assert(/^http:\/\//.test(endpoint), `Endpoint should start with 'http://', received '${endpoint}'`);
+  const self = state(endpoint);
 
   return {
     isConnected: (): boolean =>
