+++ packages/api-provider/src/http/state.spec.js
@@ -1,0 +1,13 @@
+// Copyright 2017-2018 Jaco Greeff
+// This software may be modified and distributed under the terms
+// of the ISC license. See the LICENSE file for details.
+
+const state = require('./state');
+
+describe('state', () => {
+  it('requires an http:// prefixed endpoint', () => {
+    expect(
+      () => state('ws://')
+    ).toThrow(/with 'http/);
+  });
+});
