+++ packages/api-provider/src/types.js
@@ -25,10 +25,9 @@ export type JsonRpcResponse = JsonRpcObject & JsonRpcResponseBase;
 
 export type ProviderInterface$Callback = (error: ?Error, result: mixed) => void;
 
-export interface ProviderInterface {
-  +isConnected: boolean;
-
-  send (method: string, params: Array<mixed>): Promise<mixed>;
-  subscribe (method: string, params: Array<mixed>, cb: ProviderInterface$Callback): Promise<number>;
-  unsubscribe (id: number): Promise<boolean>;
+export type ProviderInterface = {
+  isConnected (): boolean,
+  send (method: string, params: Array<mixed>): Promise<mixed>,
+  subscribe (method: string, params: Array<mixed>, cb: ProviderInterface$Callback): Promise<number>,
+  unsubscribe (id: number): Promise<boolean>
 }
